+++ src/get-generated-files/get-gitpod-dockerfile.ts
@@ -21,7 +21,6 @@ export default function () {
     RUN brew install gh
 
     # Puppeteer dependencies
-    RUN sudo apt-get update && sudo apt-get install -y libgtk-3-0 libx11-xcb1 libnss3 libxss1 libasound2 libgbm1 libxshmfence1
-
+    RUN sudo apt-get update && sudo apt-get install -y libgtk-3-0 libx11-xcb1 libnss3 libxss1 libasound2 libgbm1 libxshmfence1\n
   `;
 }