+++ src/get-generated-files/npmrc.ts
@@ -1,1 +1,1 @@
-export default 'node-linker=hoisted';
+export default { 'auto-install-peers': false, 'node-linker': 'hoisted' };